fix(react, react-vite): add 'Page' suffix
