+++ react/official/tabs/src/pages/Tab2Page.tsx
@@ -1,8 +1,9 @@
+import React from 'react';
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
 import ExploreContainer from '../components/ExploreContainer';
-import './Tab2.css';
+import './Tab2Page.css';
 
-const Tab2: React.FC = () => {
+const Tab2Page: React.FC = () => {
   return (
     <IonPage>
       <IonHeader>
@@ -22,4 +23,4 @@ const Tab2: React.FC = () => {
   );
 };
 
-export default Tab2;
+export default Tab2Page;
